layout: modify layouts to handle a xfer with N objects during a PUT
This patch changes the layouts to take into account the fact that xfers
can now contain N objects during a PUT with the no-split option.
The change only concerns the functions used when writing, in other
cases the layouts remains unchanged (read).

For this, in the pho_encoder structure, N I/O contexts and N
layout_info structures are now initialized at the beginning to contain
the information of the N objects.

During writing, we iterate over these N I/O contexts and layout_info to
write all the objects. For this, an index representing the object
being written has been added in many functions to retrieve the correct
I/O context or layout_info structure.

Change-Id: I986b4eb280ad081e58978548a80e6212244bd456
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13988
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>