cli: modify XferDescriptor and put to handle N targets in a xfer
This patch modifies the XferDescriptor to initialize N targets.
Also, the mput_register function has been added to take as input a
list of (oid, src, attrs). The mput_register function is used only if
we are doing a mput with the no-split option, otherwise the
put_register is used.

Change-Id: Ic89159f3a7b40a3138f4effcb50704ea38c95b6b
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13994
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>