layout: add delete IO context for RAID encoders
The IO context encapsulates important information for the ongoing IO.
For the delete decoder, it mainly consists in the number of extents the
operation still need to delete.

Also, allow one to distinguish a READ decoder to a DELETE decoder, using
the delete_action boolean of the pho_encoder structure.

Change-Id: Ia2f68a1b8ba9688a4295cc249f288978a52060ce
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14001
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>