store: modify hard delete behavior to use layout encoder
Before this patch, the hard delete behavior was hard coded in the store
module, mainly because only tapes were handled and no effective
deletions were done.

Now, the hard delete command uses the behavior implemented in layout
encoders. That allows the management of all families which have an
IO adapter which implements a delete operation.

A test is removed from non-regression, as the new behavior will not
manage deprecated versions of the targeted object: only its living
version. A ticket is created to be able to manage directly deprecated
objects.

Test will come in the following patch.

Change-Id: I129d8846c4278f1dc9d1353be30257f0b5dae36d
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/13934
Reviewed-by: Gauthier Evraerd <gauthier.evraerd@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>