raid: add an overhead to the write alloc to anticipate the size of xattrs and directory
This patchs adds a new value in the conf to add or not an overhead
when doing a write alloc. The overhead is used to anticipate the size
taken by the xattrs and 2 directories created for each object.
It was an issue when doing a mput --no-split with intermediate sync.
We could make an alloc of the total size of a medium but end up with
an error indicating that there is no more space in the medium.

Change-Id: I13f7d386429f8a2181a2fb2846a76c177902384c
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14051
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>