sql: move lyt_info, obj_status and *_time columns from object to copy
This patch moves the lyt_info, obj_status and *_time columns from
object and deprecated object to copy table. The creation_time is kept
in the object and deprecated object. The obj_status is renamed to
copy_status in the copy table.

The lyt_info, obj_status and *_time will be deleted from the
object_info structure when switching with the 3.0 schema.

Change-Id: Ib6b3895bcbf09ab5de7f6a483193be17192ed5b9
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14107
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>