store: add copy list functions
This patch adds in the Store API the copy list function.
The copy list is added earlier in order to keep all the tests using
the obj_status/cpy_status. The copy list function can take as input a
list of uuid and a cpy_status filter. For now, it is a simple version,
it will be improved and extended later.

Also, the status_filter is kept in the object list until we
switch from 2.2 to 3.0 schema.

Change-Id: I10505bd82a807c97d11472dc45a4522b5bb215d7
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14109
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>