cli: remove `__init__.py`
This removal is made because of future renaming of `cli.py` to
`__init__.py`. Git will not let the `git mv` command resolve without
applying a useless +3000;-3000 delta.

This change implies that `phobos` is a namespace package, until
the renaming of `cli.py`, without any differences: the
non-regression tests and the RPM installation will still work as
desired.

Change-Id: Icd55373355a01db11d9344a086a80fc34015c3ed
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14097
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Gauthier Evraerd <gauthier.evraerd@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>