sql & store: switch to 3.0 schema, create and get default copy
This patch switches to the 3.0 schema which add the copy table and
moves the lyt_info, copy_status and *_time columns from object to copy.
It's also removed them from the object_info structure.
Therefore, all the dss functions of the objects, deprecated objects
and layouts are updated to take into account the modifications.

With a put or a get, we create or retrieve the default copy of the
object. Also, everywhere we were dealing with obj_status/copy_status,
*_time or layout, we now retrieve the default copy before.

Finally, the tests are updated to take into account the new copy
table.

Change-Id: I4cd18de406d7ae181fa672f9177bc688cc695b19
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14112
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>