cli: add a test with a phobos get after several migrations
This test check that we can still do a get after doing one or several
database migrations. The test put an object with the 1.91 schema and
do a get with the latest version.

The put is simulate in the test because phobos requires a specific
schema version. When we are doing the put, phobos requires the latest
version and not the 1.91. To bypass this, we initiate the DSS with the
data corresponding to a put with the 1.91 version.

Change-Id: I8ecb1ff4dec6887e698f17b2f8068e5125313fc6
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14130
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>