layout: data_processor with a posix reader and a writer
To target copy create, we want to explicitly set a reader and a
writer into a data processor. Encoders will have a posix reader
and layout writer. Decoders will have a layout reader and a posix
writer. Copy creator will have a layout reader and a layout
writer. For eraser, we could use a reader or a writer as target
and no reader or writer as a source.

This patch first introduces reader and writer into the data
processor.

Posix reader and writer are implemented in separated files.
Following patchs will implement reader and writer for layout and
a new data processor logic to deal with reader and writer.

Change-Id: Iba05b35e52c513d556996c5507b1fc1547bbd34d
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14262
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>