dss: add wrapper function to get an object by oid,uuid,version,deprecated(-only)
This patch adds a new wrapper function to get an object. By default,
it will retrieve object only from the alive objects. Its behavior can
be extend with the deprecated and deprecated-only parameter to
retrieve also or only from the deprecated objects. With the deprecated
parameter, it will first try to retrieve from the alive objects, if
not found will search in the deprecated objects.

Also, if uuid is not provided and there are several objects found, no
object will be chosen and an error will be returned.
If the version is not provided and there are several objects found,
the object with the latest version will be chosen.

Change-Id: I7e7d02c43b2bad1c8e0be49f785c88af725cc7c7
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14267
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>