bugfix: incomplete set of orphan extents
This patch fixes an issue with the hard delete of extents on tapes. In
that case, if there were replicas, we would only set the first one as
orphan because the loop was based on the number of layout retrieved from
the database. Since that number is always 1, we wouldn't tag every
extent as orphan.

Change-Id: Id1a911e4579276c9151872c37b3a12eb3916b726
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14288
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>