store: add order when fetching copies
This patch changes the logic when fetching copies. If no copy is
specified, it will first try to get the copy specified in the
`get_preferred_order` parameter in the conf. Next, it will try to get
the default copy. If no copy has been found yet, it will get the first
copy from the dss.

Change-Id: I389b798980bb2f3f0b1ab0291ee36d9ff17b5fa6
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Signed-off-by: LUCAS Patrice <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14313
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>