cli/store: update rename with the new arguments
This patch updates the arguments for the rename command. Now, the oid
must be set. The behavior of the command remains unchanged. If only
the OID is provided, the latest version of the object must be in the
alive table. If the UUID is also provided, the latest version of the object
can be in either the deprecated or alive table.

Change-Id: Ic2378fdb995536c483001f8973f6d517f318467a
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14394
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>