cli/store: update undelete with the new arguments
This patch updates the undelete command to use the new arguments for
the object commands. Undelete now takes a list of OIDs as input. Also,
the `--uuid` option has been added. An UUID can only be used with
one oid.

If multiple objects are found, an error is returned. However, if
multiple objects are found and an UUID is provided, the object with
the latest version will be undeleted.

Change-Id: Ibd486e9924a980000f2c2e7db969404355d9e390
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14395
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>