cli: add new arguments for object list
This patch adds the `--uuid`, `--version` and `--deprecated-only`
options to the object list. The `--uuid` option can only be used with
one oid. If no oid or more than one oid is provided, an error is
returned.

Also, the `--deprecated` option, which was already present, has been
modified. It now specifies that we want to display both alive and
deprecated objects. Therefore, the `--deprecated-only` option has been
added to display only the deprecated objects.

Change-Id: I20ea28e6d13e4a8fbb18209e7bceb16ba8cb4838
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14398
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>