lrs: do not decrement the health counter without TLC
If we can't contact the TLC, we should not decrement the health
counter of the medium or drive, as this is not an error related to the
medium or drive.

Also, if we lost the TLC, the request is not requeued and error is
returned to the client. Otherwise, as the device/medium is not set to
failed, the LRS will try indefinitely to satisfy the request without
success.

Change-Id: I42d8e10a1515c71945449ba83bd21a598428a751
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14495
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>