cli: add copy create
This patch adds the exec_create function for copy operations. The
arguments for copy create are the same as those used for the put
command and only the --copy-name argument from the get command.

Currently, only one target is allowed in a xfer for a copy operation.
However, this can be extended in the future to support multiple
targets by modifying the arguments of the command.

Change-Id: I78fb4b1b7f7aaf5412331617c54a041a3902c623
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14498
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>