dss: update renew condition for DSS locks following locate
In case of an early lock in a locate command, the PID stored in the
database is the client one, which leads to fail the renew without
forcing it because the daemon gets another PID.

We now avoid checking the PID in case of an early lock.

Add two unit tests to check the behavior of the `is_early` field of a
lock, for the refresh and the unlock requests.

Change-Id: Ief4e16763637bed3fedaebab34c143590741f8f6
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14534
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>