dss: update lock insertion to consider `is_early`
The lock table insertions mainly consists of two different calls:
`dss_lock` and `dss_lock_hostname`. The first one is used by the daemon
while the second is used by clients when doing a locate operation.

In this case, only the latter needs to indicate that it inserts an early
lock.

Also some modification is the dss_lock internals, to manage generic
calls with three parameters.

Change-Id: Ic463c5cb97c6a41d7e2ea160f4b6af77873cb832
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14533
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>