autotools: add `cli` files when doing `make install`
When doing `make install`, using `phobos_db` failed because the python
packages were not installed. Now they are, and the `phobos` script also.

There is still an issue because `phobos` is installed in `/bin`, meaning
a standard user can execute it, but the python packages are not
available for a non-sudo user. This may be discussed later, when
deciding if `phobos` can work in userspace.

Change-Id: I1e46ca2247ebee53e1b3bfb390426d4acec2deab
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14586
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>