layout: use the io_block_size parameter
This patch puts back the use of the io_block_size parameter. The
io_block_size replace can now be use as the chunk size. If this
parameter is set, the stripe size is now io_block_size * n_data_extents.

Also, this patch ensures that the splits (except the last one) have a
size that is a multiple of the stripe size.

Change-Id: I4ffa616ee95fbe31d36bd87a408bef4a14cbcae5
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14749
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>