prod: use the TEST_DRIVES env variable
This patch adds the proper handling of the `TEST_DRIVES` environment
variable by the prod/acceptance tests. Its goal is to define a list of
drives to use before running the test for pre-production testing. For
non-regression tests, the current behaviour is unchanged. The syntax of
this variable is also updated to match the `TEST_TAPES` one.

Change-Id: I96f5f202bfe3e763c7f33f4e35f2f9965f47c586
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14832
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>