cfg: add a PHO_CFG_GET_SUBSTRING_VALUE macro
This macro will retrieve and parse cfg values holding information for
different families. It is different from the
`pho_cfg_get_substring_value` function as it uses a list of
`pho_config_item` to detail code-level default values, which the
function did not handle (which is a bug).

This macro will replace all calls to `pho_cfg_get_substring_value` in a
later patch.

Change-Id: I13e54da090c2afdf5622ca85a62f8281347f9629
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14847
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>