tests: fix fs_block_size retrieval and mput test
This patch fixes an issue with the retrieval of the `fs_block_size`
configuration parameter, which did not use the default values specified
in the code. Alongside this, it also fixes the mput no-split test which
was checking certain put would fail, but for the wrong reasons, as the
`fs_block_size` was not properly considered.

Change-Id: I63bfdecb07d54efe4d6b6790efcc204204298808
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14848
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>