context: remove default initialization for mock functions
This patch removes the default initialization for the mock functions, to
instead set them to NULL by default.

This allows us to remove the unnecessary include from the context file,
but more importantly, the context manager does not need to know the
actual value of these mocked functions, only the module which uses them
does.

Therefore, only it should set its value if the function is NULL.

Change-Id: Ie33a1e74bf2875767903497357a9363adbc05322
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14905
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>