mock: add mock function for mput errors after first sync
This patch adds a mock function to create an error during an mput
after the first sync succeeded.

Modify the `test_mput_sync_error` to use that new mock function. It will
currently simply return an error on the second call, and a later patch
will add the handling of the rollbacks on the database's info.

Change-Id: I829ba57c2d2513715197881cdccbe9c6929636b3
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14906
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>