misc: add format macro for `struct pho_id`
This will ease the formatting of resources ID, like devices and media,
and help keeping log consistency.

For now, modifications are only made in two DSS files. Code will be
updated progressively.

Also, change checkpatch to allow the macro definition, which asked for
parentheses around the macro, which is not possible.

Change-Id: Ib6b5c1669145557f39d8779d5bc971813c176ba7
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14867
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>