store: keep objects and copies on a partial put success
This patch changes the handling of objects, copies and extents after a
put failed after a partial synchronization. In that case, the extents
written should be kept in the database, and associated objects and
copies should be updated accordingly, instead of everything just being
erased.

To do this, we add a return code to the `pho_xfer_target` structure to
record the result of each target, instead of having just a global rc for
the whole xfer.

This field is only used for multi write requests, but could be used for
multi read or copy too.

Note: with this patch, the objects and copies are properly kept and
updated in the database, but the extents are not because they are not
written in the data_processor structure until the final release. This
will be changed in a next patch.

Change-Id: I3e019519f552aea0f3e44f8b3b7a501653166d25
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14909
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>