lrs: fix partial release race condition
The partial release is done by setting to false the ld_ongoing_io
flag to allow the sync that put back this flag to true. No other
IOs must be scheduled during this partial release. So, we need to
be sure that there is no gap between setting the ld_ongoing_io
flag to false and the time we force a sync.

Change-Id: Ie49051467b7609f30a2d0991e5d872d4f36e381f
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/14992
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>