admin: reuse tags from the source during repack
This patch adds support for reusing tags from the source medium by
default. This means that if the --tags option is not specified, Phobos
will search for another medium with the same tags as the source.

Also, if --tags "" is explicitly specified, Phobos will consider any
available medium, restoring the previous behavior when the option was
not used.

This commit fix the github issue #44.
This commit integrates the github pull request #47.

Change-Id: I5e66df99ea19e47cb85b99ec9d163b59d8343fee
Signed-off-by: Gauthier EVRAERD <gauthier.evraerd.externe@ayudata.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/15130
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>