lrs: prevent negative physical space free update
Repack, with its huge amount of newly written extents, could try
to update a negative physical space free at release. This is due
to the fact that we ignore automatic compression that happens on
magnetic tape. The correct value will be set by the statfs at
sync. Meanwhile, we set the secure 0 value.

Negative value is not accepted by dss_media_get.

This patch fix the github issue #53 .
This patch integrates the github pull request #54 .

Change-Id: I3942585c9b76e67134bf7770bf9580f9d8cef629
Signed-off-by: Patrice LUCAS <patrice.lucas@ayudata.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/15047
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>