admin: allow `load` to operate after a `locate`
Whenever a `load` operation occurs, a lock is taken over the tape to
ensure no one will use it. But if a `locate` operation happened, an
early lock is registered in the database and will prevent the lock of
the `load`.

Now, the `load` operation can manage early locks and will juste take the
ownership of the early lock.

A test is added to check this behavior.

Change-Id: Ie36462b21119e65b531f74865b02ba03eb3f2211
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/15049
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>