cli: fix test_migration_integrity for RHEL9
In PostgreSQL 13.22, a `\restrict key` field has been added to pg_dump
result. If not provided, the key is generated randomly. It was causing
the test to failed because between two pg_dump, the key was different.
This patch fixes the test by checking if the psql version is greater
or equal to 13.22. If so, a key is provided to pg_dump.

Change-Id: Iee4bac8abf970f78ae01c94df9db5a5372c44c66
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-on: https://gerrit.ccc.ocre.cea.fr/gerrit/c/phobos/+/15069
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Yoann VALERI <yoann.valeri@cea.fr>
Reviewed-by: Gauthier EVRAERD <gauthier.evraerd@cea.fr>
Tested-by: Jenkins s8open_nr <s8open_nr@ccc.ocre.cea.fr>