dss, cli: add extent creation_time management
We can now get extent creation time from DSS.

The extent creation_time at import is read and
set from stat_buf.st_ctim of the imported file
of the extent.

The extent creation_time at repack is set with the DSS
'now' default as soon as the extent is created.

The extent creation_time at PUT is set to the io adapter
close time because deffered extent insertion could lead
to an extent creation_time imprecision if we only use the
DSS default 'now' value.

Change-Id: Ibaf03c957c59c5f4cb8f7429d59a99ec14c55319
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>