store: fix bugs when I/O fails
A few bugs where seen while trying to write a test
an inconsistency between the size given in an xfer
and the actual size of a file:

- we don't set the result of the I/O operation in the xfer
  structure on failure which leads the store layer to believe
  that the I/O succeeded;
- we don't release the media on error;
- we don't properly check the object size when calling the reader's
  step function.

If a data processor is not responsible for an error, it should
not send a release with an error. For instance, if the reader
processor fails, the writer processor should release its media

This patch fixes these issues and adds a test where the size
given in the xfer is greater than the file's size to ensure
phobos will fail accordingly.

This patch fix the lustre-hsm-phobos github issue #8 .
This patch integrates the phobos github pull request #59 .

Change-Id: Ia620c15e646fd18f3f375cdc007a49d229f17cf3
Signed-off-by: Guillaume Courrier <guillaume.courrier.externe@ayudata.fr>
Signed-off-by: Patrice Lucas <patrice.lucas@cea.fr>