cli/api: tape lost management
This patch fix the github issue #31 .
This patch integrates the github pull request #54 .

This patch adds the deletion of the media declared lost from the
database, and all record of the extents on it.

This patch adds the import verb for the dir resource. The main goal is
for tests with the `dir delete` feature, to showcase that a `dir delete
-> dir import` will produce the same database content as before the
delete.

This patch adds the reconstruction function for the raid4 layout,
allowing us to determine from the layout of a raid4 copy if that copy is
still complete, readable or incomplete.

This patch adds the deletion of the layout associated with each extent
on the medium to declare lost, and the reconstruction of the associated
copy.

This patch adds the retrieval of the layout associated with each extent on
the medium to delete. Using this layout, we'll be able to get the copy
and update its readability status.

This patch adds the handling of the `lost` parameter of the media delete
function. If true, we will not skip over media containing extents, and
instead of simply using the `dss_media_delete` function, call a new
one to remove the media, the extents marked as on it, and update the
objects and copies as necessary.

This patch adds the `--lost` flag to the `delete` command for media. If
used, the command will also delete extents marked as on the media from
the database, and update the objects and copies accordingly.

Change-Id: I7b1ad696237ba88da42b4b416918ddcb7a785db6
Signed-off-by: Patrice LUCAS <patrice.lucas@cea.fr>
Signed-off-by: VALERI Yoann <yoann.valeri.externe@ayudata.fr>