dss: refactoring lock API
Basically improving the way we can interact with locks in the database,
where one needed to unlock then lock again to change the owner of a
lock. Now a refresh query is available to do the same.

The next patch will add the first replacement of the unlock/lock
behavior.

Change-Id: Ibb8cb4d4d8ce8995f0189ce7c357d76ad2321d6c
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>